# Timbangan/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Timbangan Digital IoT API",
        "description": "API untuk integrasi perangkat timbangan IoT (Arduino/ESP8266) dengan sistem penimbangan digital. Semua endpoint menggunakan token perangkat untuk autentikasi.",
        "contact": {
            "name": "Timbangan Digital Team"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": " Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/fg-psn/settings": {
            "get": {
                "tags": [
                    "FG PSN"
                ],
                "summary": "Get pengaturan sesi aktif timbangan FG PSN",
                "description": "Dipanggil oleh Arduino/ESP8266 untuk mendapatkan informasi sesi aktif operator FG PSN.",
                "operationId": "7dfbfe34cad227fad3204420e5179ff5",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token unik perangkat timbangan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "DEV-TOKEN-FG-PSN-001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SettingsResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/fg-psn/weight": {
            "post": {
                "tags": [
                    "FG PSN"
                ],
                "summary": "Kirim data berat timbangan FG PSN",
                "description": "Menerima data berat dari ESP8266/Arduino untuk modul Finished Goods PSN.",
                "operationId": "2c326db47457d234c19d24043bbdb8d8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "token",
                                    "weight"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "DEV-TOKEN-FG-PSN-001"
                                    },
                                    "weight": {
                                        "description": "Berat dalam kg",
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.25
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berat berhasil disimpan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WeightSuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Berat tidak valid / Tidak ada sesi aktif",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SettingsResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "ready",
                            "idle"
                        ],
                        "example": "ready"
                    },
                    "kode_produksi": {
                        "type": "string",
                        "example": "KP-2024-001"
                    },
                    "nama_produk": {
                        "type": "string",
                        "example": "Mie Goreng 85g"
                    },
                    "operator": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "expired": {
                        "type": "string",
                        "example": "2025-12-31"
                    }
                },
                "type": "object"
            },
            "WeightSuccessResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Data berat 25.5 kg tersimpan"
                    },
                    "record_id": {
                        "type": "integer",
                        "example": 42
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            }
        }
    },
...
}